Modify all logic to use hibernate. Little architects changing.

+++ src/main/resources/shop_schema.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS role
 (
-    id        INT PRIMARY KEY AUTO_INCREMENT,
+    id        BIGINT PRIMARY KEY AUTO_INCREMENT,
     user_role VARCHAR(50) NOT NULL
 );
 
@@ -13,13 +13,13 @@ CREATE TABLE IF NOT EXISTS users
     password     VARCHAR(50)        NOT NULL,
     phone_number VARCHAR(20) UNIQUE NOT NULL,
     address      VARCHAR(100)       NOT NULL,
-    role_id      INT,
+    role_id      BIGINT,
     FOREIGN KEY (role_id) REFERENCES role (id)
 );
 
 CREATE TABLE IF NOT EXISTS order_status
 (
-    id          INT PRIMARY KEY AUTO_INCREMENT,
+    id          BIGINT PRIMARY KEY AUTO_INCREMENT,
     status_name VARCHAR(50) NOT NULL
 );
 
@@ -29,7 +29,7 @@ CREATE TABLE IF NOT EXISTS shop_order
     customer_id         BIGINT NOT NULL,
     order_creation_date DATE   NOT NULL,
     order_close_date    DATE,
-    status_id           INT,
+    status_id           BIGINT,
     FOREIGN KEY (customer_id) REFERENCES users (id),
     FOREIGN KEY (status_id) REFERENCES order_status (id)
 );
@@ -39,7 +39,7 @@ CREATE TABLE IF NOT EXISTS product
     id                  BIGINT PRIMARY KEY AUTO_INCREMENT,
     product_name        VARCHAR(50)    NOT NULL,
     product_description VARCHAR(200)   NOT NULL,
-    price               DECIMAL(15, 2) NOT NULL,
+    price               FLOAT(15, 2)   NOT NULL,
     product_remain      INT            NOT NULL
 );
 
@@ -52,4 +52,3 @@ CREATE TABLE IF NOT EXISTS product_order
     FOREIGN KEY (order_id) REFERENCES shop_order (id),
     FOREIGN KEY (product_id) REFERENCES product (id)
 );
-
